fixed compiler warning

+++ src/remote.c
@@ -70,7 +70,7 @@ static gboolean remote_server_client_cb(GIOChannel *source, GIOCondition conditi
 		GList *work;
 		gchar *buffer = NULL;
 		GError *error = NULL;
-		guint termpos;
+		gsize termpos;
 
 		while ((status = g_io_channel_read_line(source, &buffer, NULL, &termpos, &error)) == G_IO_STATUS_NORMAL)
 			{
